Finished AddSong and Search function
