+++ AuthDemoProject/Models/Song.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace AuthDemoProject.Models
 {
     public class Song
     {
+        
         public int Id { get; set; }
 
         public string Name { get; set; }
