+++ AuthDemoProject/Views/Artist/Add.cshtml
@@ -1,4 +1,6 @@
-﻿<h1>Add New Artist</h1>
+﻿@model AuthDemoProject.ViewModels.AddArtistViewModel
+
+<h1>Add New Artist</h1>
 
 <form asp-controller="Artist" asp-action="ProcessAddArtistForm" method="post">
     <div class="form-group">
