+++ AuthDemoProject/Views/Home/AddSong.cshtml
@@ -1,4 +1,6 @@
-﻿<h1>Add A Song</h1>
+﻿@model AuthDemoProject.ViewModels.AddSongViewModel
+
+<h1>Add A Song</h1>
 
 <form asp-controller="Home" asp-action="ProcessAddSongForm" method="post">
