+++ AuthDemoProject/Views/Home/Index.cshtml
@@ -7,39 +7,10 @@
 <div class="text-center">
     <h1 class="display-4">Welcome to TunIn` Music Database</h1>
     <h2>By Jordon Peyton</h2>
-    <p><a asp-controller="Home" asp-action="Add">Add</a></p>
-    <p><a asp-controller="Search" asp-action="Index">Search</a></p>
+    <p><a asp-controller="Home" asp-action="AddSong">Add Songs</a></p>
+    <p><a asp-controller="Search" asp-action="Index">Search Songs</a></p>
     <p><a asp-controller="List" asp-action="Index">List</a></p>
 </div>
-
-
-@if (Model.Count == 0)
-{
-    <p>No songs yet!</p>
-}
-else
-{
-    <table class="table">
-        <tr>
-            <th>Song Name</th>
-            <th>Artist Name</th>
-            <th>Genre</th>
-        </tr>
-        @foreach (Song song in Model)
-        {
-            <tr>
-                <td><a asp-controller="Home" asp-action="Detail" asp-route-id="@song.Id">@song.Name</a></td>
-                <td>@song.Artist.Name</td>
-                <td>@song.Genre</td>
-            </tr>
-        }
-    </table>
-}
-
-
-
-
-
