+++ AuthDemoProject/Views/Song/Add.cshtml
@@ -1,4 +1,6 @@
-﻿<h1>Add New Song</h1>
+﻿@model AuthDemoProject.ViewModels.AddSongViewModel
+
+<h1>Add New Song</h1>
 
 <form asp-controller="Song" asp-action="ProcessAddSongForm" method="post">
     <div class="form-group">
