+++ AuthDemoProject/Views/Song/Index.cshtml
@@ -4,18 +4,23 @@
 
 @if (Model.Count == 0)
 {
-        <p>No songs listed</p>
+<p>No songs yet!</p>
 }
 else
 {
-        <div>
-            <ul>
-                @foreach(Song song in Model)
-{
-                        <li>
-                            <a as-controller="Song" asp-action="About" asp-route-id="@song.Id">@song.Name</a>
-                        </li>
-}
-            </ul>
-        </div>
+<table class="table">
+    <tr>
+        <th>Song Name</th>
+        <th>Artist Name</th>
+        <th>Genre</th>
+    </tr>
+    @foreach (Song song in Model)
+    {
+        <tr>
+            <td><a asp-controller="Home" asp-action="Detail" asp-route-id="@song.Id">@song.Name</a></td>
+            <td>@song.Artist.Name</td>
+            <td>@song.Genre</td>
+        </tr>
+    }
+</table>
 }
